Commented out failing test with reader macro #_
Test the reader macro comment with the TeamCity build, to see if the test is
just be skipped when the TeamCity build job is run.

It would be interesting to see the build identify any tests that are commented
out, so that broken test cannot be committed.

Commenting out tests like this is not an approved process.